Added files via upload
So, I just converted search to php and have it redirect to itself when a user submits information. This way I could just display the search results on the same page. If you want to test it, upload it to your gl server, and you can try it there. Valid inputs that should match and show up are 3, 3

lat 37.880218
long -81.062525

lat 39.296032
long -76.783868

I still need to change it later to have it connect to Luke's database. Also, have no idea how to use city/state/zipcode to get a region, find lat/long within it, and then display the points of interest within that region. Pretty sure that's more Ajax and Javascript though. Bare bones of it is done, but still need some work.