Added files via upload
the map is created upon initialization. it calls moveMaker from the form, on submit. I commented out all the other validation stuff so that it's just easier to work with. Ideally, this is my goal....once I get the ajax to work, I want to call my php file, return an array of markers, and then plots those markers on the map. But for now, I just want to be able to update the map with user inputted latitude/longitude -___-;